requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
openai-whisper>=20231117
pydub>=0.25.1
sounddevice>=0.4.6
streamlit>=1.33.0
sqlalchemy>=2.0.0
reportlab>=4.1.0
nltk>=3.8.1
requests>=2.31.0
python-multipart>=0.0.9
pytest>=8.0.0
pytest-cov>=5.0.0
httpx>=0.27.0
